requirements.txt
hydra-core>=1.3.0
numpy>=2.3.0
transformers>=4.56.0
timm>=1.0.20
pillow>=11.3.0
matplotlib>=3.10.6
wandb>=0.22.0
datasets>=4.1.0
